-bug#0000909: ALL_DATA_SOURCES_NODUPS breaks graphing when using "Consolidation Function" MAX -bug#0001239: Using the TOTAL_ALL_DATA_SOURCES in a graph where all item are using the MAX CF yields an invalid(empty) CDEF
The issue we discussed at SF concerning the use of GPRINT/CF is now solved. The bug introduced with early 087* releases is removed and the processing is mostly reverted to 086* behaviour.
The "guessing" of CF, introduced in 087*, is still available for corner cases

Verified with totals, TOTAL_*_DATA_SOURCES_*DUPS, graphs using AREA/LINEs for AVERAGE/MAX and stuff.

The planned new feature to be more flexible with GPRINTs will at least be postponed to 088; better to a version were VDEF support is introduced

git-svn-id: svn+ssh://repo.cacti.net/var/data/svnroot/cacti/cacti/branches/0.8.7@4795 860744bd-22fc-0310-8c96-e9fe5004b5ca