correct form rendering logic
if a sub-form has the same variable name as a form, the previous value
is taken from the sub-form's session variable and not the intended input
value when returning to the main form.  This is reproducable by going to
data_input.php, editing a method, going to the sub-form, which includes
the form variable 'name'.  Upon saving and returning to the main edit
form the value of the sub-form 'name' variable is injected and not the
intended value..