remove use of deprecated function get_request_var_request().  replace it with get_request_var()
this is part three of project to remove the use of $_GET, $_REQUEST, and
$_POST in the code.  Parts one and two have already been completed.
Part one was to remove the use of $_GET throughout the code, part two
was the incorporation of wrapper functions to handle specific
validation, verification and sanitization cases.  This third part will
remove the use of the get_request_var_request() from the code.  Part
four, and beyond will include the removal of both $_REQUEST and $_POST
super globals from the code.