big commit back porting features
back port the following features from main:

1) vdef support
2) new graph items
3) new graph options

this covers all the high points available in rrdtool today.

also change the following:

1) allow forms to include spacers
2) reorganize graph template form with spacers
3) many help wording back ports from main

for those testing, if you are going to start using this version, you
will have to implement the changes in the upgrade script or reinstall.
also note that this change may break new aggregates.