Attempt at fixing step and heartbeat during import
Step and heartbeat was always set to default when importing profiles with a newly created datasource_profile(example 1 minute polling profile).

Now step and heartbeat is read from the profile.

New installations and upgrades do not have a datasource_profile. $profile_id is returned as an array in that case. Not to happy about the fix for that, but it works.

Tested import with default profile, 1 minute polling profile as well as new installation import of Disk IO. Works fine.