Update the automation devices with matched host template
Probably should change this column to an int, and insert the host
template id instead of the name.  The original intent of this column was
to discover the OS for a device even without a host template to match
against.  Not sure if I want to continue down that route.