resolve login redirect loops
Corner case where you have forced a password change, but have not
granted that permission to the user.  We are also now blocking the save
when the user tries to make this setting.

We need better interactions with the login screen under certain failure
modes.  But for now this is good enough.