Resolving Issue #819
Allow saving customized timespan for Graph page.  This functionality was
removed accidentally from Cacti 1.x.  For people who have upgraded
already, there is a function that finds what is the best timespan.  If
this is not acceptable, and the Data Source Profiles are Read only, you
may have to hand alter the tables to enter the 'preferred' Timespan.