Add CPU Total to 'SNMP - Get Processor Information' (#1110)
* Add CPU Total to 'SNMP - Get Processor Information'

This change adds a virtual index 'Total' to the SNMP info,
showing the average CPU utilization over all cores.

* Changing CPU Total in 'SNMP - Get Processor Information'

Making the total CPU utilization the first element in the array to avoid shifting when CPUs are added.

* Fix for issue #1111 - Issue with ldap authenication (#1118)

* Update change log for pull request #1111

* Fixes for issues #1084, #1106 and #1116 (#1117)

* Fixes for issues 1106 (Graph filters), 1084 (Tree Branch population), and 1116 (Device Template won't set to none)

* Updated changelog\nfor issues #1084, #1106 and #1116

* Cleanup after pull request #1177

Cleanup includes some misqueues in the pull request as well as some
other miscelaneous changes associated with how we have landed on
formatting.

Chagne include:
* No soft tabs on initial tabbing
* ON clause of JOIN should be capitalized and on separate line
* AND/OR should be capitalized in SQL Queries
* Formatting of db_..() functions calls
* Use of sizeof() instead of isset() for array checks
* i18n of Secpass check return values
* Remove trailing whitespace from lines

- ToDo: Remove more having in place of WHERE clauses

* Fix issue #965 - LDAP search error results (#1112)

* Fix issue #965 - LDAP search erorr results

Corrected the error number returned by "Unable to set referrrals" to the correct number.

Added a new error for finding a users DN but not finding the user from the DN.

* Updated changelog for issue #965

* Fix issue 1106 - Graph Filters: template filter have empty value (#1114)

he filter uses index 0 as "All Graphs and Templates" but then searches graph_local for matches against graph_template_id. However, if index 0 is passed as a variable, it is also used in the search. Now when I looked at my database, the following entry existed:

select * from graph_local where graph_template_id = 0;
+------+-------------------+---------+---------------+---------------------+------------+
| id   | graph_template_id | host_id | snmp_query_id | snmp_query_graph_id | snmp_index |
+------+-------------------+---------+---------------+---------------------+------------+
| 1837 |                 0 |       0 |             0 |                   0 |            |
+------+-------------------+---------+---------------+---------------------+------------+

Therefore, it displays an index, but can't get a name for the template since there is no template with ID 0.

select * from graph_templates where id = 0;
Empty set (0.00 sec)

Added code to check for a zero value and only search the databsae if a value is present.

* Fix issue 1109 - Opening graph links by unlogged user - validation error (#1115)

* Fix issue 1109 - Opening graph links by unlogged user - validation error

Referer variable needs to be checked to ensure that it doesn't already have parameters applied.  If it does, then it needs to use the & rather than ? to join onto the URL.

* Updated changelog for issue #1109

* Fix spelling issue

* Minor updates to pull request 1109.

* Resolving Issue #1104

Different Timestamps in Cacti log when in Developer Debug

* Fix issue #1119 - ss_host_cpu.php pulling back too much data (#1120)

* As per the discussion on #1110 it would appear that ss_host_cpu is pulling back everything from the hrProcessorTable rather than just the processor usage. There appears to be a field named 'hrProcessorFrwID' being defined in the database which is actually the index field not the firmware ID

* Added CHANGELOG for issue #1119

* Fix for issue #1111 - Issue with ldap authenication (#1118)

* Update change log for pull request #1111

* Fixes for issues #1084, #1106 and #1116 (#1117)

* Fixes for issues 1106 (Graph filters), 1084 (Tree Branch population), and 1116 (Device Template won't set to none)

* Updated changelog\nfor issues #1084, #1106 and #1116

* Cleanup after pull request #1177

Cleanup includes some misqueues in the pull request as well as some
other miscelaneous changes associated with how we have landed on
formatting.

Chagne include:
* No soft tabs on initial tabbing
* ON clause of JOIN should be capitalized and on separate line
* AND/OR should be capitalized in SQL Queries
* Formatting of db_..() functions calls
* Use of sizeof() instead of isset() for array checks
* i18n of Secpass check return values
* Remove trailing whitespace from lines

- ToDo: Remove more having in place of WHERE clauses

* Fix issue #965 - LDAP search error results (#1112)

* Fix issue #965 - LDAP search erorr results

Corrected the error number returned by "Unable to set referrrals" to the correct number.

Added a new error for finding a users DN but not finding the user from the DN.

* Updated changelog for issue #965

* Fix issue 1106 - Graph Filters: template filter have empty value (#1114)

he filter uses index 0 as "All Graphs and Templates" but then searches graph_local for matches against graph_template_id. However, if index 0 is passed as a variable, it is also used in the search. Now when I looked at my database, the following entry existed:

select * from graph_local where graph_template_id = 0;
+------+-------------------+---------+---------------+---------------------+------------+
| id   | graph_template_id | host_id | snmp_query_id | snmp_query_graph_id | snmp_index |
+------+-------------------+---------+---------------+---------------------+------------+
| 1837 |                 0 |       0 |             0 |                   0 |            |
+------+-------------------+---------+---------------+---------------------+------------+

Therefore, it displays an index, but can't get a name for the template since there is no template with ID 0.

select * from graph_templates where id = 0;
Empty set (0.00 sec)

Added code to check for a zero value and only search the databsae if a value is present.

* Fix issue 1109 - Opening graph links by unlogged user - validation error (#1115)

* Fix issue 1109 - Opening graph links by unlogged user - validation error

Referer variable needs to be checked to ensure that it doesn't already have parameters applied.  If it does, then it needs to use the & rather than ? to join onto the URL.

* Updated changelog for issue #1109

* Fix spelling issue

* Minor updates to pull request 1109.

* Resolving Issue #1104

Different Timestamps in Cacti log when in Developer Debug

* Fix issue #1119 - ss_host_cpu.php pulling back too much data (#1120)

* As per the discussion on #1110 it would appear that ss_host_cpu is pulling back everything from the hrProcessorTable rather than just the processor usage. There appears to be a field named 'hrProcessorFrwID' being defined in the database which is actually the index field not the firmware ID

* Added CHANGELOG for issue #1119

* Add CPU Total to 'SNMP - Get Processor Information'

This change adds a virtual index 'Total' to the SNMP info,
showing the average CPU utilization over all cores.

* Changing CPU Total in 'SNMP - Get Processor Information'

Making the total CPU utilization the first element in the array to avoid shifting when CPUs are added.

* Update CHANGELOG