Make the snmpagent more consistent
The snmpagent plugin labeled snmp_password as snmp_auth_password and
snmp_priv_passphrase as snmp_priv_password.  This change aligns the
column names in the database between features.

Additionally, for most of Cacti, with the inclusion of noAuthNoPriv, the
snmp_auth_protocol needed to be expanded to 6 characters from 5.