Fix for Issue #1260 - Missing TTF functions (#1270)
* Fix for Issue #1260 - Missing TTF functions
This change corrects an issue where TTF is not enabled even though GD has been enabled as described in #1260.  The
code will now also list the state of TTF functionality during installation under the optional modules.

Testing can be done by modifying the either of the two function_exists() calls to an invalid function name.  If TTF is not
available, the code will attempt to place the text within the Classic tab image.  It will reduce the font size, and failing that
look for any available spaces, then trim back to this.  If all spaces have been used, it will start reducing characters one at a
time to find a fit against any of the system fonts.

* Update CHANGELOG