Resolving Issue #1263
The utility includes the options:
--report, --alters, --report, --repair, and --load (for developers)

The load will create the default tables from the current schema.  The
developer who runs with the --load option is still responsible for
loading the 'revised' table contents into the two string variables
though.  The --repair option will repair all tables and would be
suitable for the installers purposes.