Ensure compression for package extraction is correct
This is for the initial import case.  Based upon QA, we have found that
if the php base level compression is different than the way the package
were created in the first place, that they are unreadable.  Therefore,
this change ensures that all compression settings are consistent across
all operations including the yet unpublished package plugin.