Issue #1530: Inconsistent behaviour handling blank Field Name/Value when editing data query suggested values (#1533)
As reporting in #1530, when trying to add/edit a suggested value, multiple issues could occur.  A blank field name was allowed which would result in a suggestion that was never used when editing the data source suggestions.

Additionaly, when editing a graph suggestion, no output was returned because the code was not properly handling the parameter variations.