Make Top Tabs use Ajax Callback
We can almost collapse all headers into a single call now.  There is a
currently a problem with the Dark and Modern themes the first time you
switch to preview.  On the first page.  This was a bit of a hack since
the cactiContent object was not being parsed correctly by jQuery.