Feature #1677 - Default Tree node sorting to be inherited (#1678)
* Feature #1677 - Default Tree node sorting to be inherited

This feature changes the default sorting option to be 'Inherited' when creating new nodes (including root ones) whilst editing a tree. This does not update existing ones to be inherited.

Note: if a DB has already been updated to 1.2, either downgrade and upgrade again or modify the sort_children_type default on the graph_tree_items table to be 0 rather than 1.  Failure to do so will result in manual sorting still being applied do to SQL/PHP assuming that a zero means no value!

* Minor change to update script.