Issue #1674 - Threads and Processes values not migrated to Poller table during upgrade (#1679)
When upgrading to v1.2.0 a new set of fields are added to the Poller table to allow customisation of the number of processes/threads per poller.  That helps to vary the amount used if any given poller has more or less power than another.

However, the system-wide defaults should be migrated against each poller present in the poller table during the upgrade, otherwise both values default back to 1 which seriously increases the time to poll (especially if you have a super high timeout value).