Resolving issue #1651
This PR corrects an issue with the template import functionality where
browsers will appear to be attempting to upload a file but Cacti never
receives the variables to know that a file was attempted.

System temporary folder must be writable to upload a file so Cacti will
now error should the temporary folder be unwritable.