Import: Fixed handling of references to objects not included in file (#1796)
* Import: Fixed handling of references to objects not included in file

When exporting an object with the "Include Dependencies" option disabled and then importing the resulting file, the import process will not complete successfully. This is because the mapping of object hashes to object IDs only tracks objects that have been included in the file.

This commit resolves the issue by first loading the mapping with all existing hashes and IDs before processing the file's objects. This results in the expected behavior when a referenced object is not in the file but is already loaded into the Cacti installation.

* Revised #1787 based on review from @cigamit

- Removed unnecessary double quotes
- Encapsulated `foreach` on possibly-empty array with size check
- Added whitespace

* Updated CHANGELOG to include #1796