Feature #1806 - Control of low level debug (#1807)
This patch adds functionality to control the debugging of low level routines
via defines within the config.php file.  In order to allow flexibility, if
these are defined, variables within the $config array are primed as true,
otherwise false.

These elements within the $config array are defined in uppercase to highlight
the difference between these are other ordinary configuration optios.

Should a function or program need to enable debugging of these various
options, simply setting the value to true in the $config array will enable
logging from that point.  Inversely, setting it to false will stop logging
from that point.

The only option that is of limited use is DEBUG_READ_CONFIG_OPTION_DB_OPEN
that will only work up to the database being open.  Setting this to true
after the database has already been opened will not automatic turn off the
companion option DEBUG_READ_CONFIG_OPTION since the db_connect_real() will
not be called again.

The list of options and their functionality are:

- DEBUG_READ_CONFIG_OPTION : record use of read_config_option()
  to cacti-option.log in the system temporary folder

- DEBUG_READ_CONFIG_OPTION_DB_OPEN : Once the database has been
  successfully opened, disable DEBUG_READ_CONFIG_OPTION

- DEBUG_SQL_CMD : Always write a copy of the SQL being requested
  to the cacti-sql.log file in the system temporary folder

- DEBUG_SQL_FLOW : Write flow of db_xxx calls to the cacti-sql.log
  in the system temporary folder.

The system temporary folder is normally set to one of the following:

- Linux, CLI: /tmp/
- Linux, WEB: Apache/Ngix's temporary folder
- Windows, CLI: %TEMP%
- Windows, WEB: Temp folder of web user or %WINDIR%\TEMP