Fix setting graph titles using add_graphs.php
add_graphs.php supports a --graph-title argument to set the title on an
existing graph or the graph being created.  Somewhere between 0.8.8f
and 1.1.38, this broke.  The current code sets the title_cache column
in graph_templates_graphs, which is then immediately overwritten by the
call to update_graph_title_cache().  The correct action is to update
the title, not the title_cache.