Resolving issue #2384
If current authentication method is set to None, change to Builtin.

To achieve this, the code will look for a valid administration user that will be used to reset the password.  This is achieved by searching for the Primary Admin, and then any admin with console access.  Once an admin has been found, the must_change_password and change_password flags are set, the password is cleared and the user is automatically logged in before being forced to change the empty password.

If no admin can be located, the script will die to prevent issues and allow manual fixing of the problem.  The simpliest way to do this would be to utilise the INSERT statement that is used for the 'admin' user in the cacti.sql's user_auth table, plus the INSERT for user_id 1 into the user_auth_realm table.