Resolving issue #2419
Host state time was not correctly calculated if snmp_sysUptimeInstance was used.  The lib/functions.php was dividing the value by 100 but the SQL statement was not, resulting in invalid sort orders.

Additionally, the get_timeinstate() function was updated to utilise the 'instate' element if it exists in the $hosts variable.