Make the local poller timezone configurable
Some cloud SQL providers do not make the textual default_time_zone
configurable. While a manual offset value (e.g. -04:00) can be
configured for this value, that's not too convenient in places that
observe DST.

By exposing the poller timezone configuration for all pollers, it's
possible to override this default with a textual zone name if desired.