Initial commit of cli automation
This bring a first version of a CLI script which allows to apply
automation rules to devices. This comes in handy for large(ish)
installations that additionally make use of numerous automation rules
(and possibly not all of them being optimized), where using this through
the web interface often results in timeouts and/or incomplete runs, or
would require crazy high max_execution_time settings.

This currently simply wraps around automation_update_device(), for which
logging is available through cacti_log() - no output will be sent to
stdout.