Resolving Issue #2962
This should correct the boost and dsstat's issues.  There were a few hands on this one and it had not undergone a full regression in a while.

This update should speed things up pretty signifcantly.

First, don't check the last_update time if rrdtool is >= 1.5
Second, we can accurately assume that the filemtime() is the last update time.  This speeds things up dramatically.