Small update to #2977 if running RRDtool < 1.5
This change will allow RRDtool to return the lastupdate time of the RRDfile if you are using RRDtool version 1.4 or 1.3, or if using a newer version of RRDtool, but the RRDtool version is not set correctly in Cacti.

Better to simply upgrade to a more recent RRDtool as the different in performance will be pretty high.

I may may one more change to bring back some old code in order to reduce the impact on performance.