Fix issues when config.php is unreadable
This commit adds a notice when configphp is present but unreadable.  It also adds the ability to display any database connection errors that occur within db_connect_real for diagnostic purposes.  These notices are only presented to the end user when the appropriate flag is set as most users do not need them.

Closes #3181 and closes #3182