Partial resolution to low lighthouse score for performance #3402
* Cacti scores low on performance audit on lighthouse audit
* This change enables generic caching of pages, but with a very short age for RRDtool generated graphics.
* It increses the SEO score through the use of specialty tags on anchors leveraging the _blank target attribute value
* It removes and unnecessary tooltip on the document element

There may be additional changes forthcoming.  However, these changes raised the score from 0 to the mid twenties.

Some of the changes, like minification of JavaScript we do not do due to the nature of the project, but could be done by customers.  Once the page is first rendered, it's in the browsers memory and cached for a good time, so this is not a big deal.