When sorting data debug checks by user, no results are shown (#3713)
When sorting data debug checks by user, no results are shown and
errors recorded into the cacti log because the user field was being
looked up seperately.  However, there is no need for this as it can
all be done in the SQL statement so removed the extra code and had
SQL bring the user in allowing it to be sortable.

Closes #3713