Update html_utility.php to fix sorting issue
This change allows for sort to work properly when sorting on the Automations-> Discovered Devices IP address column. The INET_ATON was not getting applied when you were sorting just by the IP address column. It would work when first going to the page after sorting by IP but if attempting to sort while staying on the page it would drop the INET_ATON(ip) to be just `ip`. 

I attempted to follow the other places where this code looks to have been run to come up with a similar solution.