Getting ready for PHP Doc
This change clears out a number of useless comments in the mainline section of the code.  These will be replate by PHP Doc template comments in the near future.

This is the first directory.  The lib directory has a bit more robustness to the documentation and will take a bit more time to complete.