Fixing Issues #645 - Check Data Sources before Graph Create
So, this is not 100% finished yet as it does not currently work with the Remote Agent.  However needed some testers.
This change also deprecates Script Server calls being called without the proc_open() command which was a requirement from the early 2000's for a PHP version I don't even remember.
The change also improves logging when creating graphs via automation.
Lastly, marking #3572 as resolved as it was resolved in another fix over the weekend.