add log verbosity
When sending data to main, this happens in portions and this uses
several DB-queries. For optimization purposes, we want to know how much
time these queries took.
It is also good to know, how fast the recovery is processing data, hence
we also log the max timestamp for the current loop's data portion.