correction of recovery core loop
The previous handling was a bit off. Though this was technically not
critical, it was not correctly calculating the data sizes, which would
lead to wrong annotations in the logs.

Introducing the helper variable $sql_size helps to keep track properly
about the size of the SQL string for the currently processed record.
Previously the size was already added even if the current record was not
taken into account in the current loop run, and after transfering the
complete data-frame, the size was nulled, therefore the next size was
not reflecting this record.

Adding $record_count allows to log the actual amount of data that was
transfered to the main poller in the current loop cycle.