Poller recovery - related to #4215: log-verbosity, fixing a minor bug, code cleanup (#4224)
* removed unused variable "$end_count"

* renamed $count to more speaking $record_size

* renamed $inserted to $records_inserted

* code clean-up

* add log verbosity

When sending data to main, this happens in portions and this uses
several DB-queries. For optimization purposes, we want to know how much
time these queries took.
It is also good to know, how fast the recovery is processing data, hence
we also log the max timestamp for the current loop's data portion.

* correction of recovery core loop

The previous handling was a bit off. Though this was technically not
critical, it was not correctly calculating the data sizes, which would
lead to wrong annotations in the logs.

Introducing the helper variable $sql_size helps to keep track properly
about the size of the SQL string for the currently processed record.
Previously the size was already added even if the current record was not
taken into account in the current loop run, and after transfering the
complete data-frame, the size was nulled, therefore the next size was
not reflecting this record.

Adding $record_count allows to log the actual amount of data that was
transfered to the main poller in the current loop cycle.

Co-authored-by: Bernd <berni@berni.lan>