Correcting issues related to #4396
Moving this to a common function.  We still need discussion on when it's best to perform table maintenance for tables like the `poller_output` table, that when this table is InnoDB, it can become fragmented, though empty and that free space can not be released without an optimize operation, which for InnoDB tables can take some time.