finish rewiring rebuild_poller_cache.php
this now uses all the refactored code in lib/utility.php. The principal
control flow is now equivalent, or close to push_out_host().

As with the old code, this does not update the poller cache
data-source-by-data-source, but it one call to
poller_update_poller_cache_from_buffer(). While this is expected to
faster, it means that we cannot get any progress from it. Generally, the
call should complete quickly, but it's something to keep in mind for the
future.

Some (which?) data does get updated in the data source loop, but I'm
unfortunately far from familiar with the database
schema/architecture/design of cacti to build an reliable mental model
and hence cannot describe these workings satisfactory.