Device Description is not consistent in Poller Cache view (#4651)
Device Descriptions are normally clickable throughout the system to link back to
the device in question. However, when viewing the poller cache this isn't the
case. Whilst you can filter by device, if you are searching by data source name,
you aren't necessarily working with a singular device. This is not usually a
problem with other screens as you simply click the device description and are
sent to the device in question, but not when viewing the poller cache.