Another update for verifying remote data collectors are online
This check goes deeper into all cases of device update.  This is a quick fix mind you.  I think we need to redo this after the 1.2.20 release as there seems to be quite a bit of redundant code that I'm not entirely happy with at the moment.  But it should work for now.