improve SQL query to search for nodes in tree view (#4813)
* improve SQL query to search for nodes in tree view

Removes one of the joins on the host table: by site_id. Logic behind
that join is unclear, and on first glance would return just the whole of
graph_tree_items. This behaviour may have gone unnoticed because a
default install of cacti comes without any sites defined (not even a
default one) and site_id may contain NULL. On the install where we
observed issue #4806, sites are defined however.

* add changelog entry

* Light correction to SQL

Co-authored-by: rb83 <>
Co-authored-by: TheWitness <thewitness@cacti.net>