QA Bug fixes and features (#5146, #5148, #5149)
This corrects issues with the host edit form's availabity settings

Feature: enable/disable menu options

This introduces a new feature of disabling and/or enabling of
menu options with an option function to run a value check after
options have been updated.

Feature #5148: theme-based includes with master location

This introduces a new feature of specifying a path, theme and
file to be included.  This allows multiple locations to be
checked for the existence of a file, including the original path
or a subfolder of a theme location (see get_theme_paths).

Feature #5149: Async CSS loading

This introduces a new feature of allowing CSS files to be
loaded async by the browser.  By specifying the CSS files are
for print media only, the browser will delay loading them
speeding up initial load.  Using the javascript 'onload'
event, we then switch them back to being for all media

Closes #5148, #5149