Additional Fixes for #5336 - Poller Output Issues
This change modifies boost processing slightly in the case of a mismatch between the graph template and the data template.  In addition, it more closely aligns with the stock poller_output processing in some of the functions.

At some point, we should merge the three methods that are in place for processing poller items.