Changed error for an invalid password to match error for an invalid username. (#6483)
Prior to this change, when logging in with an invalid username, Cacti would return "Access Denied! Login Failed!" However, if the username is valid, Cacti would instead return "Access Denied! Login failed." This change capitalizes the word "failed" when using a valid username, which makes it more difficult for an attacker to enumerate accounts with a brute force attack.